requirements.txt
Django==1.8
Markdown==2.6.7
django-filter==0.15.0
djangorestframework==3.4.7
pillow==3.3.1
